# scripts/.contracts/openapi_current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fasal Saathi Backend",
    "version": "2.0"
  },
  "paths": {
...
    "/api/yield": {
      "get": {
        "tags": [
          "ML Prediction"
        ],
        "summary": "Predict Get",
        "operationId": "predict_get_api_yield_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ML Prediction"
        ],
        "summary": "Predict Yield",
        "description": "Yield prediction using ML router",
        "operationId": "predict_yield_api_yield_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/notifications": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get Notifications",
        "operationId": "get_notifications_api_notifications_notifications_get",
        "parameters": [
          {
            "name": "crop",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Crop"
            }
          },
          {
            "name": "irrigation_count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Irrigation Count"
            }
          },
          {
            "name": "water_coverage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "title": "Water Coverage"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PredictRequest": {
        "properties": {
          "Crop": {
            "type": "string",
            "maxLength": 50,
            "title": "Crop"
          },
          "CropCoveredArea": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cropcoveredarea"
          },
          "CHeight": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cheight"
          },
          "CNext": {
            "type": "string",
            "maxLength": 50,
            "title": "Cnext"
          },
          "CLast": {
            "type": "string",
            "maxLength": 50,
            "title": "Clast"
          },
          "CTransp": {
            "type": "string",
            "maxLength": 50,
            "title": "Ctransp"
          },
          "IrriType": {
            "type": "string",
            "maxLength": 50,
            "title": "Irritype"
          },
          "IrriSource": {
            "type": "string",
            "maxLength": 50,
            "title": "Irrisource"
          },
          "IrriCount": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Irricount"
          },
          "WaterCov": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Watercov"
          },
          "Season": {
            "type": "string",
            "maxLength": 50,
            "title": "Season"
          }
        },
        "type": "object",
        "required": [
          "Crop",
          "CropCoveredArea",
          "CHeight",
          "CNext",
          "CLast",
          "CTransp",
          "IrriType",
          "IrriSource",
          "IrriCount",
          "WaterCov",
          "Season"
        ],
        "title": "PredictRequest"
      },
      "PredictResponse": {
        "properties": {
          "predicted_ExpYield": {
            "type": "number",
            "title": "Predicted Expyield"
          }
        },
        "type": "object",
        "required": [
          "predicted_ExpYield"
        ],
        "title": "PredictResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}